Sprint 8. Fix tests. Abstract Task Manager Test.
