+++ test/service/InMemoryTaskManagerTest.java
@@ -22,28 +22,6 @@ class InMemoryTaskManagerTest extends TaskManagerTest<InMemoryTaskManager> {
     }
 
     @Test
-    @DisplayName("Create Task")
-    public void shouldCreateTaskAndEqualsToHimself() {
-        Task task1 = new Task(task.getName(), task.getDescription(), LocalDateTime.now().plusMinutes(55),
-                Duration.ofMinutes(15));
-        task1.setId(task.getId());
-        assertEquals(task1, task, "Task should equal to his copy");
-        taskManager.createTask(task1);
-        assertNotNull(taskManager.getTask(task1.getId()));
-    }
-
-    @Test
-    @DisplayName("Create Epic")
-    public void shouldCreateEpicAndEqualsToHimself() {
-        Epic epic1 = new Epic(epic.getName(), epic.getDescription());
-        epic1.setId(epic.getId());
-        epic1.addTask(subTask.getId());
-        assertEquals(epic1, epic, "Epic should equal to his copy");
-        taskManager.createEpic(epic1);
-        assertNotNull(taskManager.getEpic(epic1.getId()));
-    }
-
-    @Test
     @DisplayName("All Subtasks Status NEW")
     public void shouldAllSubtasksStatusNew() {
         assertTrue(taskManager.getSubTasks().stream()
@@ -83,94 +61,6 @@ class InMemoryTaskManagerTest extends TaskManagerTest<InMemoryTaskManager> {
         assertTrue(taskManager.getSubTasks().stream()
                 .allMatch(subTask -> Status.IN_PROGRESS.equals(subTask.getStatus())));
         assertEquals(Status.IN_PROGRESS, epic.getStatus());
-    }
-
-    @Test
-    @DisplayName("Create SubTask")
-    public void shouldCreateSubTaskAndEqualsToHimself() {
-        SubTask subTask1 = new SubTask(subTask.getName(), subTask.getDescription(), subTask.getEpicId(),
-                LocalDateTime.now().plusMinutes(90), Duration.ofMinutes(15));
-        subTask1.setId(subTask.getId());
-        assertEquals(subTask1, subTask, "SubTask should equal to his copy");
-        taskManager.createTask(subTask1);
-        assertNotNull(taskManager.getTask(subTask1.getId()));
-    }
-
-    @Test
-    @DisplayName("Delete tasks one by one")
-    public void shouldDeleteOneByOne() {
-        taskManager.deleteByIdTask(task.getId());
-        assertNull(taskManager.getTask(task.getId()));
-        assertEquals(taskManager.getTasks().size(), 0);
-
-        taskManager.deleteByIdSubTask(subTask.getId());
-        assertNull(taskManager.getTask(subTask.getId()));
-        assertEquals(taskManager.getSubTasks().size(), 0);
-
-        taskManager.deleteByIdEpic(epic.getId());
-        assertNull(taskManager.getTask(epic.getId()));
-        assertEquals(taskManager.getEpics().size(), 0);
-    }
-
-    @Test
-    @DisplayName("Delete all tasks")
-    public void shouldDeleteAll() {
-        taskManager.deleteAllTasks();
-        taskManager.deleteAllEpics();
-        taskManager.deleteAllSubTasks();
-        assertEquals(taskManager.getTasks().size(), 0);
-        assertEquals(taskManager.getEpics().size(), 0);
-        assertEquals(taskManager.getSubTasks().size(), 0);
-    }
-
-    @Test
-    @DisplayName("Update task")
-    public void shouldUpdateTask() {
-        assertEquals(task, taskManager.getTask(task.getId()), "Task should equal to his copy");
-        task.setName("nameNew");
-        task.setDescription("descriptionNew");
-        task.setStatus(Status.DONE);
-        taskManager.updateTask(task);
-        assertEquals(task, taskManager.getTask(task.getId()), "Task should equal to his new copy");
-    }
-
-    @Test
-    @DisplayName("Update epic")
-    public void shouldUpdateEpic() {
-        assertEquals(epic, taskManager.getEpic(epic.getId()), "Epic should equal to his copy");
-        epic.setName("nameNew");
-        epic.setDescription("descriptionNew");
-        taskManager.updateEpic(epic);
-        assertEquals(epic, taskManager.getEpic(epic.getId()), "Epic should equal to his new copy");
-    }
-
-    @Test
-    @DisplayName("Update subTask")
-    public void shouldUpdateSubTask() {
-        assertEquals(subTask, taskManager.getSubTask(subTask.getId()), "SubTask should equal to his copy");
-        subTask.setName("nameNew");
-        subTask.setDescription("descriptionNew");
-        subTask.setStatus(Status.DONE);
-        taskManager.updateSubTask(subTask);
-        assertEquals(subTask, taskManager.getSubTask(subTask.getId()), "SubTask should equal to his new copy");
-    }
-
-    @Test
-    @DisplayName("Get task")
-    public void shouldGetTask() {
-        assertNotNull(taskManager.getTask(task.getId()));
-    }
-
-    @Test
-    @DisplayName("Get epic")
-    public void shouldGetEpic() {
-        assertNotNull(taskManager.getEpic(epic.getId()));
-    }
-
-    @Test
-    @DisplayName("Get task")
-    public void shouldGetSubTask() {
-        assertNotNull(taskManager.getSubTask(subTask.getId()));
     }
 
     @Test
@@ -235,16 +125,5 @@ class InMemoryTaskManagerTest extends TaskManagerTest<InMemoryTaskManager> {
                     .plusMinutes(50),
                     Duration.ofMinutes(15)));
         });
-    }
-
-    @Test
-    @DisplayName("Prioritized Tasks should be in order")
-    public void shouldPrioritizeTasks() {
-        Task taskNew = taskManager.createTask(new Task("name4", "description4", LocalDateTime.now()
-                .minusMinutes(60),
-                Duration.ofMinutes(15)));
-        SubTask subTaskNew = taskManager.createSubTask(new SubTask("name5", "description5",
-                epic.getId(), LocalDateTime.now().plusMinutes(15), Duration.ofMinutes(10)));
-        assertEquals(List.of(taskNew, task, subTaskNew, subTask), taskManager.getPrioritizedTasks());
     }
 }
